Fix stuff around to track the progress of a Share. File.cpp:24 still need to be fixed to have a correct value

+++ src/share.cpp
@@ -31,9 +31,9 @@ float Share::progress() const
     uint total = size_total();
 
     if(total == 0)
-        return 0;
+        return 0.0;
 
-    return size_local() / size_total();
+    return (float)size_local() / (float)size_total();
 }
 
 uint Share::size_total() const
